chore(Key): use tsx
